Handle the status action
