+++ src/backend/api/BasicAuth.ts
@@ -1,5 +1,7 @@
+import {AuthResultType} from 'alinea/cloud/AuthResult'
 import {atob} from 'alinea/core/util/Encoding'
 import {Auth, AuthedContext} from '../Backend.js'
+import {AuthAction} from '../Handler.js'
 
 export function basicAuth(
   verify: (username: string, password: string) => boolean | Promise<boolean>
@@ -14,20 +16,19 @@ export function basicAuth(
   }
   return {
     async authenticate(ctx, request) {
-      const auth = request.headers.get('Authorization')
-      if (!auth) return unauthorized()
-      const [scheme, token] = auth.split(' ', 2)
-      if (scheme !== 'Basic') throw unauthorized()
-      const [username, password] = atob(token).split(':')
-      const authorized = await verify(username, password)
-      if (!authorized) return unauthorized()
-      const from = new URL(request.url)
-      return new Response('Authorized', {
-        status: 302,
-        headers: {
-          location: from.pathname
+      const verified = await this.verify(ctx, request)
+      const url = new URL(request.url)
+      const action = url.searchParams.get('auth')
+      switch (action) {
+        case AuthAction.Status: {
+          return Response.json({
+            type: AuthResultType.Authenticated,
+            user: verified.user
+          })
         }
-      })
+        default:
+          return new Response('Bad request', {status: 400})
+      }
     },
     async verify(ctx, request): Promise<AuthedContext> {
       const auth = request.headers.get('Authorization')
